Extract pipe logic
